[Chore] 사진 아이템 매퍼 이동
