+++ app/src/main/java/com/and04/naturealbum/ui/maps/MapScreenViewModel.kt
@@ -4,14 +4,8 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.and04.naturealbum.data.dto.FirebaseFriend
-import com.and04.naturealbum.data.dto.FirebaseLabelResponse
-import com.and04.naturealbum.data.dto.FirebasePhotoInfoResponse
 import com.and04.naturealbum.data.repository.DataRepository
 import com.and04.naturealbum.data.repository.FireBaseRepository
-import com.and04.naturealbum.data.room.Label
-import com.and04.naturealbum.data.room.PhotoDetail
-import com.and04.naturealbum.utils.toLocalDateTime
-import com.naver.maps.geometry.LatLng
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -68,34 +62,5 @@ class MapScreenViewModel @Inject constructor(
                 Log.d("FriendViewModel", e.toString())
             }
         }
-    }
-}
-
-// Room Data -> UI Data
-fun Label.toLabelItem() = LabelItem(name, backgroundColor)
-fun List<PhotoDetail>.toPhotoItems(labels: List<Label>): List<PhotoItem> {
-    val labelMap = labels.associate { roomLabel -> roomLabel.id to roomLabel.toLabelItem() }
-    return map { photoDetail ->
-        PhotoItem(
-            photoDetail.photoUri,
-            LatLng(photoDetail.latitude, photoDetail.longitude),
-            labelMap.getValue(photoDetail.labelId),
-            photoDetail.datetime
-        )
-    }
-}
-
-// FireBase Data -> UI Data
-fun FirebaseLabelResponse.toLabelItem() = LabelItem(labelName, backgroundColor)
-fun List<FirebasePhotoInfoResponse>.toFriendPhotoItems(labels: List<FirebaseLabelResponse>): List<PhotoItem> {
-    val labelMap =
-        labels.associate { firebaseLabel -> firebaseLabel.labelName to firebaseLabel.toLabelItem() }
-    return map { firebasePhotoInfo ->
-        PhotoItem(
-            firebasePhotoInfo.uri,
-            LatLng(firebasePhotoInfo.latitude!!, firebasePhotoInfo.longitude!!),
-            labelMap.getValue(firebasePhotoInfo.label),
-            firebasePhotoInfo.datetime.toLocalDateTime()
-        )
     }
 }
