+++ app/src/main/java/com/and04/naturealbum/ui/maps/PhotoKey.kt
@@ -1,5 +1,10 @@
 package com.and04.naturealbum.ui.maps
 
+import com.and04.naturealbum.data.dto.FirebaseLabelResponse
+import com.and04.naturealbum.data.dto.FirebasePhotoInfoResponse
+import com.and04.naturealbum.data.room.Label
+import com.and04.naturealbum.data.room.PhotoDetail
+import com.and04.naturealbum.utils.toLocalDateTime
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.clustering.ClusteringKey
 import java.time.LocalDateTime
@@ -29,3 +34,32 @@ data class PhotoItem(
     val label: LabelItem,
     val time: LocalDateTime,
 )
+
+// Room Data -> UI Data
+fun Label.toLabelItem() = LabelItem(name, backgroundColor)
+fun List<PhotoDetail>.toPhotoItems(labels: List<Label>): List<PhotoItem> {
+    val labelMap = labels.associate { roomLabel -> roomLabel.id to roomLabel.toLabelItem() }
+    return map { photoDetail ->
+        PhotoItem(
+            photoDetail.photoUri,
+            LatLng(photoDetail.latitude, photoDetail.longitude),
+            labelMap.getValue(photoDetail.labelId),
+            photoDetail.datetime
+        )
+    }
+}
+
+// FireBase Data -> UI Data
+fun FirebaseLabelResponse.toLabelItem() = LabelItem(labelName, backgroundColor)
+fun List<FirebasePhotoInfoResponse>.toFriendPhotoItems(labels: List<FirebaseLabelResponse>): List<PhotoItem> {
+    val labelMap =
+        labels.associate { firebaseLabel -> firebaseLabel.labelName to firebaseLabel.toLabelItem() }
+    return map { firebasePhotoInfo ->
+        PhotoItem(
+            firebasePhotoInfo.uri,
+            LatLng(firebasePhotoInfo.latitude!!, firebasePhotoInfo.longitude!!),
+            labelMap.getValue(firebasePhotoInfo.label),
+            firebasePhotoInfo.datetime.toLocalDateTime()
+        )
+    }
+}
